add error handling / skip checks
much better checks to make sure we only process items that have passwords!
Also, always save the report even on a crash, and lastly
not parsing url anymore, just putting whatever
is on the 1pass file and truncated to 50 chars just in case for long urls.
Moved URL field to last column